+++ plugins/sonarqube/src/components/page/Page.tsx
@@ -6,14 +6,14 @@ import {
   PageSectionVariants,
   Title,
 } from '@patternfly/react-core';
-import React, { useEffect, useState } from 'react';
+import React, { useMemo, useState } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 
 import { IOptions } from '../../utils/interfaces';
 import { IPluginPageProps } from '@kobsio/plugin-core';
 import PageToolbar from './PageToolbar';
 import Projects from './Projects';
-import { getOptionsFromSearch } from '../../utils/helpers';
+import { getInitialOptions } from '../../utils/helpers';
 
 const Page: React.FunctionComponent<IPluginPageProps> = ({
   name,
@@ -23,7 +23,7 @@ const Page: React.FunctionComponent<IPluginPageProps> = ({
 }: IPluginPageProps) => {
   const location = useLocation();
   const history = useHistory();
-  const [pageOptions, setPageOptions] = useState<IOptions>(getOptionsFromSearch(location.search));
+  const [pageOptions, setPageOptions] = useState<IOptions>(useMemo<IOptions>(() => getInitialOptions(), []));
 
   // changePageOptions is used to change the options to get a list of projects from SonarQube. Instead of directly
   // modifying the options state we change the URL parameters.
@@ -32,13 +32,9 @@ const Page: React.FunctionComponent<IPluginPageProps> = ({
       pathname: location.pathname,
       search: `?query=${encodeURIComponent(opts.query)}`,
     });
-  };
 
-  // useEffect is used to set the options every time the search location for the current URL changes. The URL is changed
-  // via the changePageOptions function. When the search location is changed we modify the options state.
-  useEffect(() => {
-    setPageOptions(getOptionsFromSearch(location.search));
-  }, [location.search]);
+    setPageOptions(opts);
+  };
 
   return (
     <React.Fragment>
@@ -47,7 +43,7 @@ const Page: React.FunctionComponent<IPluginPageProps> = ({
           {displayName}
         </Title>
         <p>{description}</p>
-        <PageToolbar name={name} query={pageOptions.query} setOptions={changePageOptions} />
+        <PageToolbar name={name} options={pageOptions} setOptions={changePageOptions} />
       </PageSection>
 
       <Drawer isExpanded={false}>