+++ plugins/sql/src/components/page/Page.tsx
@@ -6,33 +6,30 @@ import {
   PageSectionVariants,
   Title,
 } from '@patternfly/react-core';
-import React, { useEffect, useState } from 'react';
+import React, { useMemo, useState } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 
+import { IOptions } from '../../utils/interfaces';
 import { IPluginPageProps } from '@kobsio/plugin-core';
 import PageSQL from './PageSQL';
 import PageToolbar from './PageToolbar';
-import { getQueryFromSearch } from '../../utils/helpers';
+import { getInitialOptions } from '../../utils/helpers';
 
 const Page: React.FunctionComponent<IPluginPageProps> = ({ name, displayName, description }: IPluginPageProps) => {
   const location = useLocation();
   const history = useHistory();
-  const [query, setQuery] = useState<string>(getQueryFromSearch(location.search));
+  const [options, setOptions] = useState<IOptions>(useMemo<IOptions>(() => getInitialOptions(), []));
 
   // changeOptions is used to change the options for an SQL query. Instead of directly modifying the options state we
   // change the URL parameters.
-  const changeOptions = (q: string): void => {
+  const changeOptions = (opts: IOptions): void => {
     history.push({
       pathname: location.pathname,
-      search: `?query=${q}`,
+      search: `?query=${opts.query}`,
     });
-  };
 
-  // useEffect is used to set the options every time the search location for the current URL changes. The URL is changed
-  // via the changeOptions function. When the search location is changed we modify the options state.
-  useEffect(() => {
-    setQuery(getQueryFromSearch(location.search));
-  }, [location.search]);
+    setOptions(opts);
+  };
 
   return (
     <React.Fragment>
@@ -41,14 +38,14 @@ const Page: React.FunctionComponent<IPluginPageProps> = ({ name, displayName, de
           {displayName}
         </Title>
         <p>{description}</p>
-        <PageToolbar query={query} setQuery={changeOptions} />
+        <PageToolbar options={options} setOptions={changeOptions} />
       </PageSection>
 
       <Drawer isExpanded={false}>
         <DrawerContent panelContent={undefined}>
           <DrawerContentBody>
             <PageSection style={{ minHeight: '100%' }} variant={PageSectionVariants.default}>
-              {query.length > 0 && <PageSQL name={name} query={query} />}
+              {options.query.length > 0 && <PageSQL name={name} query={options.query} />}
             </PageSection>
           </DrawerContentBody>
         </DrawerContent>