+++ plugins/sql/src/utils/helpers.ts
@@ -1,8 +1,13 @@
-// getQueryFromSearch is used to get the sql query from a given search location.
-export const getQueryFromSearch = (search: string): string => {
-  const params = new URLSearchParams(search);
+import { IOptions } from './interfaces';
+
+// getInitialOptions is used to get the initial options from the url.
+export const getInitialOptions = (): IOptions => {
+  const params = new URLSearchParams(window.location.search);
   const query = params.get('query');
-  return query ? query : '';
+
+  return {
+    query: query ? query : '',
+  };
 };
 
 export const renderCellValue = (value: string | number | string[] | number[]): string => {